requirements.txt
# Core AI/ML frameworks
langgraph>=0.3.0
langchain>=0.3.0
langchain-community>=0.3.0
langchain-anthropic>=0.3.0
langchain-openai>=0.2.0
deepagents>=0.2.0

# Vector databases
chromadb>=0.5.0
sentence-transformers>=2.2.0

# Document processing
pypdf>=5.0.0
python-docx>=1.0.0
textract>=1.6.5
python-magic>=0.4.27

# API and web
requests>=2.31.0
httpx>=0.27.0
python-dotenv>=1.0.0
pydantic>=2.0.0
pydantic-settings>=2.0.0

# Data processing
pandas>=2.0.0
numpy>=1.24.0

# Utilities
tqdm>=4.66.0
python-dateutil>=2.8.2

# Development
pytest>=7.4.0
pytest-asyncio>=0.21.0
black>=23.0.0
ruff>=0.1.0
